Docker and HomeBrew improvements
HomeBrew being a rolling release, it is not straightforward to know which packages versions were used to build the third party libraries.
So we fix a dedicated commit (the 4.0.2 openmpi upgrade) and use it to install our software in brew.
For GEOSX to know it, we define this BREW_HASH as a metadata of the tarball of the TPL uploaded in the bucket. GEOSX can therefore know which hash to install against.

Two main docker images are used in GEOSX: centos7.5 (with specific LLNL compilers) and ubuntu18.04 in a more standard way. Each image exists in two flavors (clang version 6 or 7, gcc version 7 or 8).

I’ve decided to rebuild the docker images from the very beginning every time the TPL repository is modified.
I believe this will make the process more readable and easier to maintain.
For centos7.5 we spend additional 5min20 per build and 2 more minutes for ubuntu18.04.
As a benefit, we do not have to worry building the very initial base images anymore.
And only one Dockerfile is enough to build each of our images.

The Dockerfiles make use of multi-stage buildings to get rid of temporaries and to provide to GEOSX the very minimum set of tools it needs. This results in lighter images, but most important, this may help preventing additional implicit dependencies that may arise without notice.
As part of the docker contract, each docker image defines a GEOSX_TPL_DIR environment variable pointing to  the installation directory of the TPL.
GEOSX does not need to know this location anymore, it just has to use this variable.

An additional dockerBuilds repository also exists, but it is deprecated.
This patch merges the last elements in TPL we could therefore remove dockerBuilds.

This patch suggest another organization of the docker repositories and tags.
There now only exists one repository per flavor (2 flavors per os means 4 images in total).
Then we use tags to make the distinction in between the TPL builds.
On advantage can also be the cleaning, dockerhub allowing bulk cleaning for tags…

For the moment, we use the unique TRAVIS_BUILD_NUMBER as a tag for both OSX tarball and docker images.
But this may be improved, taking into account the PR, date, hash…

Last, the definition of a dedicated build command for trilinos can ease building and computer with limited memory.
And having pt-scotch part of the tplMirror will make things easier when building on an environment without internet.